Splits Refinements to own component
