+++ app/data/searchkit.json
@@ -1,0 +1,34 @@
+{
+  "connection": {
+    "host": "https://elastic.wgbh-mla.org",
+    "apiKey": "X3NoUXlJMEJZNE9yTDhJMHdMSEQ6N1RLcDQxYm9USEdCV1ByeXJ4MXFDUQ=="
+  },
+  "search_settings": {
+    "search_attributes": [
+      { "field": "title", "weight": 3 },
+      "exhibits_exhibitpage__body_edgengrams",
+      "ov_collections_collection__introduction_edgengrams",
+      "featured"
+    ],
+    "result_attributes": [
+      "title",
+      "exhibits_exhibitpage__body_edgengrams",
+      "ov_collections_collection__introduction_edgengrams"
+    ],
+    "highlight_attributes": ["title"],
+    "snippet_attributes": [
+      "exhibits_exhibitpage__body_edgengrams",
+      "ov_collections_collection__introduction_edgengrams"
+    ],
+    "facet_attributes": [
+      { "attribute": "content_type", "field": "content_type", "type": "string" }
+    ],
+    "filter_attributes": [
+      {
+        "attribute": "featured",
+        "field": "exhibits_exhibitpage__featured_filter",
+        "type": "string"
+      }
+    ]
+  }
+}
